fix(dashboard): :bug: fix disabled btn pull novel & add validation exists novels
