+++ packages/ui/src/icons/toast/index.ts
@@ -1,0 +1,2 @@
+export * from "./solar-shield-warning-bold-duotone";
+export * from "./solar-info-circle-bold-duotone";
