+++ packages/ui/src/icons/toast/solar-info-circle-bold-duotone.tsx
@@ -1,0 +1,24 @@
+import React from "react";
+import type { SVGProps } from "react";
+
+export function SolarInfoCircleBoldDuotone(props: SVGProps<SVGSVGElement>) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="1em"
+      height="1em"
+      viewBox="0 0 24 24"
+      {...props}
+    >
+      <path
+        fill="currentColor"
+        d="M22 12c0 5.523-4.477 10-10 10S2 17.523 2 12S6.477 2 12 2s10 4.477 10 10"
+        opacity={0.5}
+      ></path>
+      <path
+        fill="currentColor"
+        d="M12 17.75a.75.75 0 0 0 .75-.75v-6a.75.75 0 0 0-1.5 0v6c0 .414.336.75.75.75M12 7a1 1 0 1 1 0 2a1 1 0 0 1 0-2"
+      ></path>
+    </svg>
+  );
+}
